Enforce system certificates updates on Windows
The stack tool could not download the root.json file from hackage. It failed
with a TLS exception caused by a missing CA for the server certificate.

It seems that opening a web page with Internet Explorer or with powershell
causes the certificates to be updated.

See https://github.com/commercialhaskell/stack/issues/234