Add rules_haskell module version 0.17 (#1029)
* Add rules_haskell module version 0.17

* Add patch fixing stack_snapshot extension usage

* Add patch to remove obsolete aliases referring to nixpkgs core module

* Build only //tools/... and //haskell/...

* Set BAZEL_USE_CPP_ONLY_TOOLCHAIN

* Install libtinfo5 and libgmp-dev on Debian / Ubuntu

* Only build / test //test/... from test module

* Enforce system certificates updates on Windows

The stack tool could not download the root.json file from hackage. It failed
with a TLS exception caused by a missing CA for the server certificate.

It seems that opening a web page with Internet Explorer or with powershell
causes the certificates to be updated.

See https://github.com/commercialhaskell/stack/issues/234

* Enable cc toolchain resolution

* Add patch for windows

Register the cc toolchain that comes with the GHC bindist.

* Remove bcr_test_module for now

Our test module depends on rules_nixpkgs' modules which are not available yet.

* Set BAZEL_DO_NOT_DETECT_CPP_TOOLCHAIN=1 on Windows

* Skip Windows on CI

It is currently failing with:
```
this rule is missing dependency declarations for the following files included by 'gzwrite.c':
  'C:/b/f3c3gu33/execroot/_main/external/rules_haskell~0.17~haskell_toolchains~rules_haskell_ghc_windows_amd64/mingw/x86_64-w64-mingw32/include/stdio.h'
  'C:/b/f3c3gu33/execroot/_main/external/rules_haskell~0.17~haskell_toolchains~rules_haskell_ghc_windows_amd64/mingw/x86_64-w64-mingw32/include/corecrt_stdio_config.h'
  'C:/b/f3c3gu33/execroot/_main/external/rules_haskell~0.17~haskell_toolchains~rules_haskell_ghc_windows_amd64/mingw/x86_64-w64-mingw32/include/corecrt.h'
...
```