Add pigz@2.8
This is mostly a copy-upgrade from 2.7 that brings the module and its
dependencies up to date. Yet, I decided to replace the pigz `cc_library`
by a `cc_binary` as that is what pigz actually is and that also makes
the binary usable via the label "@pigz". For backwards compatibility, I
kept an alias "bin".

Signed-off-by: Matthias Maennich <maennich@google.com>