liburing@2.5: build against (default) libc (#1502)
The default `./configure` invocation sets up the build for static
linking (skipping any std library dependencies). That causes this module
to be broken as is. It builds, but it can't be linked against as the
built-in libc replacement functionality from src/nolibc.c is not
compiled along. Fix this by following distributions like Debian and
Fedora, and build the package against the supplied libc by specifying
`--use-libc`.

It is unlikely that somebody already used this module as it was broken
and only merged a few days ago.

Signed-off-by: Matthias Maennich <maennich@google.com>
Co-authored-by: Yun Peng <pcloudy@google.com>