Add rules_jsonnet
Even though we haven't performed a release of rules_jsonnet in ages,
it's still seeing some active use. Adding it to BCR is relatively easy,
because it mostly just depends on jsonnet and jsonnet_go. The
documentation parts depend on skydoc and rules_sass, but those aren't
necessary to use the rules themselves.