Fix c-blosc2 (#1773)
https://github.com/AcademySoftwareFoundation/openexr/pull/1604 revealed
that the c-blosc2 module still has some linker issues. This PR fixes
those issues. This PR is an improvement of the c-blosc2 model at version
2.12.0.

In detail, missing c files were added to the c-blosc2 target. Also,
textual_hdrs were reworked. Besides this cc_tests targets were
introduced to make sure the whole thing works across different systems.

The defines `HAVE_PLUGINS` and `HAVE_ZSTD` were added. Without the
define `HAVE_PLUGINS` zfp plugin is not initialized (which is required
by OpenEXR).

@phaedon Maybe this fixes also some issues on your side ;). You can
create a PR and add yourself as a maintainer to this module if you like.
This way you will get notified if "improvements" are made or at least it
gives you a chance to veto changes ;). Furthermore, the PR policy was
changed in a way that a positive review of a maintainer can lead under
some circumstances to a successful merge without the need to get a
review of one of the "admins/repo owners"