Add glm@1.0.0.bcr.1 which fixes glm@1.0.0. (#1848)
Before this change, running `bazel query "labels(hdrs, @glm)"` resulted
in `INFO: Empty results`. With this change the module correctly exports
the public headers of glm.

The differences are:

* Removed `strip_prefix` from `source.json`, the top level directory in
the zip must be retained. This fixes the main issue, the module is
unusable without this.
* Separated public and private headers into `hdrs` and `srcs`
respectively. IIUC `textual_hdrs` is meant for public headers as well,
whereas the `*.inl` files (and files in the `detail/` sub directory) of
glm are not meant to be included directly.