Add curl with curl_config.h
This reworks the curl BUILD file to use curl's cmake generated header
file. This requires replacing all possible options, but makes sure we
can't miss any important defines. For example in the previous build the
CA cert chain of trust was invalid because the define for the system
location for certs was not provided. Now we should no longer be able to
miss settings like that when curl adds new ones.