Update to new versor version - old version unusable (yanked) (#2219)
The BUILD file of the old version assumed different file paths like
`include`, `include/vsr` and `include/vsr/detail`. They are not present
in the old version.

This PR updates to a new version of versor and yanks the old. There are
no dependents yet. So, this should be safe.