feat: add `ape@1.0.0-beta.8`
Switches to assimilating the APE binaries.

This release adds a new `ape_assimilate` rule that takes a "fat" APE binary
and outputs a native binary.

The assimilated binary is entirely hermetic.

This avoids the need for runfiles when starting the binary with the APE
launcher.

Avoiding the need for runfiles removes a bootstrapping problem: to find
the runfiles in Shell, one needs _at least_ `readlink` and/or other POSIX
tools.

As `ape` is going to be bottom of the stack in lots of cases, this is
unacceptable.

Assimilating the binaries avoids the need for any runfiles.