Add freertos@11.1.0 + RP2040 support (#2635)
Hi there! Please consider this pull request to add FreeRTOS v11.1.0 to
BCR.

This updates the BUILD.bazel file slightly, adding the RP2040 port of
FreeRTOS into the build. I have verified that this builds and correctly
flashes an RP2040 device from Pigweed. Unfortunately, my Pigweed repo
has some modifications that have not been pushed yet, in part because it
depends on this change. So, you will not be able to test this yourself.
Please accept instead this lovely screenshot of a Pigweed console
pointing at an RP2040 running a version of Pigweed built on this new
version:


![image](https://github.com/user-attachments/assets/554877db-b114-4d61-8677-4e4bc947643c)

Beside the modification to BUILD.bazel, I also had to do funky things in
portmacro.h, due to the "define-static-as-empty-token" shenanigans in
the tasks_c target.