spdlog@1.15.0.bcr.2 (#3425)
Compared to `1.15.0.bcr.1`, add a `bool_flag` for header_only mode and
now depends on `@bazel_skylib`.

This enables `--@spdlog//:header_only=False` on the bazel command line.

---

See also #1769 which requests a mechanism to enable `USE_STD_FORMAT`. If
desirable, I could try adding that option as an additional flag in this
pull request, or in a follow-up change.