Added version 0.2.2 of foxglove_schemas_protobuf, added py_pb2 target. (#3551)
This PR bumps up the version of the foxglove_schemas_protobuf module to
version 0.2.2. This is based off of version 0.2.1.bcr.1 with the
addition of a consolidated py_pb2 target (it cannot be itemized due to a
bug in py_proto_library, see
https://github.com/protocolbuffers/protobuf/issues/17889).

For further info, see release notes on v0.2.2:

https://github.com/foxglove/schemas/releases/tag/releases%2Fpython%2Ffoxglove-schemas-protobuf%2Fv0.2.2

I tested this on our own project.

@bazel-io skip_check unstable_url

There does not appear to be a "stable" url, just following what was
already there in previous versions.