Fix symlink issue with boost.asio
Fixes https://github.com/bazelbuild/bazel-central-registry/issues/3631

It seems there is a problem with symlinking a directory in the BCR. The
fix in this case is to create all the sub-directories, and symlink each
individual file to the file in the previous version.

It could be worth adding a presubmit check for symlinks to directories,
to ensure this doesn't happen again in the future.

I would also recommend adding ubuntu2004_arm64 to the presubmit
platforms wherever possible, for all packages.