bcr_validation: check if a commit/reference in the github url is from the original repo (#3630)
A source archive URL in the form of
`https//github.com/OWNER/REPO/archive/<commit>.zip` isn't guaranteed to
point source code from the original repository because `<commit>` can be
a fork and GitHub will happily generate the source archive under this
URL. An attacker can use such an URL to confuse BCR maintainer that the
URL is from the source repo.

This PR adds a check to make sure the commit is actually coming from the
original repository via a GitHub API (unofficial).

---------

Co-authored-by: Xùdōng Yáng <wyverald@gmail.com>
Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>