Add asio-grpc@2.6.0
This is possible to add now that the issues with rules_cc and protobuf
have been resolved. Previously you could not have both boost and grpc
as dependencies because boost would pull rules_cc up to 0.1.0 and grpc
has a transitive dep (protoc-gen-validate) that was not comptible with
rules_cc 0.1.0.