Add facebook/folly@2025.01.13
Closes https://github.com/bazelbuild/bazel-central-registry/issues/3557

Depends on https://github.com/bazelbuild/bazel-central-registry/pull/3750

The BUILD.bazel files here are based off of the BUCK files from folly.
BUCK2 is a very similar build system, so migrating was more tedious than
difficult.

There are quite a few tests that I disabled for now. I'm working through
some of them still.

There are also some targets that are commented out for now due to
external deps that are not yet in the BCR. There are clear TODO's next
to each of these. The remaining ones are either optional / minor / or
only affect one or two targets. I would be ok with merging this without
the missing deps, and re-enabling the disabled targets via a bcr.1
release if the missing deps are added to the BCR in the future.