Add imgui 1.91.8 for Linux, MacOS, Windows (#3795)
Note that ImGui offers a matrix of `{Platform, Renderer}` backends:

https://github.com/ocornut/imgui/blob/master/docs/BACKENDS.md#list-of-backends

In this MR I have only added a target to support `{GLFW, OpenGL3}` which
works well on standard MacOS / Linux setups.
This target is named `"@imgui//:imgui-glfw-opengl3"`

and I'm assuming that if other folks are interested in extending this,
they might create similarly named targets, such as
```
        "@imgui//:imgui-glfw-vulkan",
        "@imgui//:imgui-win32-dx11",
```
etc. For now it seems like a good start to keep the BUILD rules
relatively simple and to just expose one target to clients.
Another possible model would be to decouple the targets, something like

```
        "@imgui//:imgui-core",
        "@imgui//:imgui-platform-glfw",
        "@imgui//:imgui-renderer-opengl3",

        "@imgui//:imgui-platform-win32" # etc etc
        "@imgui//:imgui-renderer-dx11",
```
but this would require the client to explicitly list 3 targets. Let me
know if you have any thoughts on this.

---------

Co-authored-by: Fædon Jóhannes Sinis <fs@treble.tech>
Co-authored-by: Ezekiel Warren <zekewarren@gmail.com>