Added `libunwind`. (#3872)
Fixes #3655.

cc @kgreenek

Added `1.8.1`, which is the latest release version and matches what
`vcpkg` has.

I wasn't able to get the `ptrace`, `coredump`, and `setjmp` tests to
pass, which means I probably didn't do the `cc_library`'s for those
features correctly... but this should be good as a first pass. We can
iterate.