add jemalloc@5.3.0 (#3896)
Initial support for the latest stable release of jemalloc is
implemented.

The autoconf/make build is complex but every attempt has been made to be
faithful to it while also following Bazel idioms. The compile time
feature detection has been shifted to static build constraints and flags
many of which depend on details of libc, the target abi, etc. Optional
build flags are exposed and could be set on the target platform or more
likely through a convenience rule that creates the build transition.
Other lower level details such as the page size are expected to be
declared by the target platform. Recognizing that this creates DX
friction, an attempt is made out of the box to choose intelligent
defaults based on the most common os/arch settings.

I have not tested windows or freebsd, though I imagine that someone with
enough motivation could either working fairly quickly. An effort was
made to transcribe feature detection wherever possible.


ref #3757