Fixed how `librdkafka` files can be included in other C/C++ files. (#3954)
Fixed `librdkafka` so it can be included as either:

```c
#include <rdkafka.h>
```

or

```c
#include <librdkafka/rdkafka.h>
```

In order to do this I needed to add `strip_include_prefix` and add an
`include_prefix` to the `librdkafka` lib. Also needed to split the lib
into a `cpp` version and a `c` version due to it having two different
source directories.