Verify github user id in metadata.json (#4191)
Whenever a module is added or its metadata is changed, verify the github
handle of the module maintainer matches the previously recorded github
user id. This is to prevent security risk in case of module maintainer's
github username change or deletion which will make the previous username
available for take over.

---------

Co-authored-by: Xùdōng Yáng <wyverald@gmail.com>