Update `folly` to build with `jemalloc`, `libunwind`, `libdwarf`, and `boost.crc` now that they are all on the BCR (#4119)
Folly is large, the easiest way to review this is to view the diff
generated by the `Generate module diff / generate_module_diff
(pull_request)` action.

But here's a summary of changes:

* Enabled building with `jemalloc`, `libunwind`, `libdwarf` (had to
patch some source files to make sure it includes the right Dwarf), and
`boost.crc`.
*   Added a Bazel flag to enable building with `int128`.
* Fixed some `folly/debugging/exception_tracer` targets by setting
`alwayslink = True`.
*   Some minor test target changes.

Also to improve the Bazelfying experience for Folly:

*   Added a script that syncs from BCR Folly to a Folly repo.
*   Updated the README with more detailed instructions.