Fix assimp build with --dynamic_mode off (#4724)
Fixes linker issues when building with `--dynamic_mode off` that come
from missing symbols in the IO plugin libraries. The fix is to merge all
IO plugin sources directly into the `@assimp` target instead of building
them as separate cc libraries. This change is backward compatible since
the plugin libraries were not publicly visible. For the sake of logical
organization, I have kept the sources for each IO plugin as separate
`filegroup` targets that are merged into `srcs` for the `@assimp`
library target.

Signed-off-by: Shameek Ganguly <shameek@intrinsic.ai>