Fix stb_image symbols ODR issue for assimp (#4749)
Linking fails for downstream target if it depends on both assimp and
another library that embeds `stb_image` due to ODR violations.

The following fixes are needed:
- Remove `STB_USE_HUNTER` from `defines` attribute for
`@assimp//:common`: Assimp prefixes the `stbi_` methods with `assimp_`
with this change so that the `stb_image` symbols are renamed in the
output library.
- Remove `STB_IMAGE_IMPLEMENTATION` from `local_defines` attribute for
`@assimp`: This centralizes the `stbi_` method implementations in the
Assimp.cpp translation unit.
- Remove `ASSIMP_USE_STB_IMAGE_STATIC` from `defines` attribute for
`@assimp//:common`: This flag has to be removed for the `assimp_`
prefixed `stb_image` symbols to be visible to targets outside of the
Assimp.cpp translation unit.

Signed-off-by: Shameek Ganguly <shameek@intrinsic.ai>