Add aom (#5063)
Note: this is my first BCR contribution.

I am not quite sure whether the module as well as the top-level build
target should be called `aom` or `libaom`. It seems to be referenced as
only aom at https://aomedia.googlesource.com/aom/, but it builds
`libaom.so` and I've seen that used in other places.