Enforce that `compatibility_level` is monontically increasing (#5165)
This makes it easier to reason about the relation between version and
compatibility level, especially when using `max_compatibility_level`.

The docs already suggest this property, but don't state it as an
invariant. We can hold the BCR to this standard while we decide whether
this should become a hard requirement for all registries.