Fix Boost date time 1.88.0.bcr.1 (#5512)
This PR fixes the visibility of Boost `date_time@1.88.0.bcr.1`. I know
there is a guideline and good reasons why existing modules should not be
changed.

Nevertheless, without this fix, it would be necessary to introduce
`1.88.0.bcr.2` - This will force me to introduce also
`boost_pin@1.88.0.bcr.2`, and this in turn will force me to update all
80 libs that are currently referenced by `boost_pin@1.88.0.bcr.1`. I
have some scripting on my side to automate this, but I think this will
lead to many non-sense boost versions, until I get it right.

A user can currently not depend on the `date_time@1.88.0.bcr.1` module,
since nothing is visible from outside.

The do not use "nodep" deps feature is hard to handle for libs such
Boost.

@fmeum @meteorcloudy @Wyverald @kotlaja