Fix boost@1.88.0.bcr.1 compatibility level (#5585)
WARNING: This PR modifies existing modules.

This PR fixes the `compatibility_level` for Boost 1.88.0. For some
unknown reason for Boost 1.88.0 the compatiblity level `108800` and
`1108800` has been used. It should be `108800` to fit into the scheme -
at least it should not differ:

| Boost version | compatibility_level |
|---------------|---------------------|
| 1.83.0        | 108300              |
| 1.83.0.bcr.1  | 108300              |
| 1.87.0        | 108700              |
| 1.88.0    | 108800              |
| 1.88.0.bcr.1  | 108800              |

First discoverd this issue with the compatible_level in
https://github.com/bazelbuild/bazel-central-registry/pull/5565

@fmeum @Wyverald @kotlaja @meteorcloudy