Correct Bzlmod tool's handling of gazelle directives
migrate_to_bzlmod.py was erroring when encountering gazelle directives
in go_repository calls in WORKSPACE files because of a backlash in an
f-string interporlation.

Moving the interpolation out to its own variable fixes the issue. We add
a new go_repository with a gazelle directive to test this case.

Fixes #5655