[New-Module] Publish grpc_kotlin v1.5.0-rc.1 to BCR. (#5639)
Hi

We are from https://github.com/grpc/grpc-kotlin. We would like to
publish it to BCR.

I followed the guidelines and create this PR. Please let us know if
there is any more details we need to complete. Also would be great if
you can let us know if there is something we need to do about
`grpc_kotlin.20250824-191240.json` which fails the tests since it's
outside `modules/` but it's generated by the tool

Thanks a lot
Thuan