feat(blake3): Add version 1.8.2 with optional oneTBB support (#5717)
BLAKE3 version 1.8.2

Adds oneTBB Support - This can be enabled with the
`--define=blake3:tbb=enabled` flag. I've added a `README.md` that
explains this and a new presubmit task.

I have continued the overlay structure that was used in `1.5.4.bcr.1`
instead of using patches.

The existing presubmit matrix and the specific CI configuration for
Windows ARM64 have been preserved from the previous version as I lack a
local Windows environment to validate any changes. I have also kept the
`bazel_compatibility` of the prior version.

This is my first contribution to the BCR. I am very open to all feedback
and happy to make any necessary changes. I have also added myself as a
maintainer in `metadata.json`. Please let me know if this is acceptable
or if you'd prefer a different approach.

Signed-off-by: James Wilson <206078155+daemoninstitute@users.noreply.github.com>