fix(lz4): v1.10.0.bcr.1 - restore targets for compatibility (#5922)
1.10.0 included an overly aggressive refactoring of the `BUILD.bazel`
file into a single unified `:lz4` target, which was an unintentional
breaking change for the following modules:
- c-blosc-2
- flann
- folly
- libarchive
- librdkafka
- mcap
- squashfs-tools

when they were dependencies in a build and 1.10.0 was also included (as
seen in issue #5899), overriding their specified dependence on earlier
versions of LZ4 that including different targets.

1.10.0.bcr.1 reverts the root `BUILD.bazel` file to the exact structure
used in 1.9.4.bcr.2. This resolves the visibility errors encountered
when building modules that depended on the `@lz4//:lz4`, `@lz4//:lz4_hc`
and `@lz4//:lz4_frame` targets.

The new multithreading feature in the `programs/` directory is
preserved.

Apologies for originally overlooking this.

Signed-off-by: James Wilson <206078155+daemoninstitute@users.noreply.github.com>