Update CI to ignore yanked versions for compatibility checks (#6322)
This is in response to
https://github.com/bazelbuild/bazel-central-registry/pull/6179 where an
incorrect compatibility version was added to a boost module and now
future versions must carry forward this value. Instead, the module with
the incorrect version should have that version yanked and a new `bcr.*`
version put in front of it. As long as yanked versions are considered,
this problem can never be fixed.