fix(bcr_validation): expand symlink check to regular files (#6357)
Github Pages doesn't support symlinks, and preprocessing needs to be
done for forks that use it to serve the BCR. There was a discussion in
issue #4192 that led to adding a check, but that check was only
performed on directory symlinks.

As per the conversation, a typical use is to symlink
`overlay/MODULE.bazel` to `../MODULE.bazel`, which are non-directory
symlinks. This extends the symlink check to regular files as well.

--

The main rationale for this change is that Github Pages doesn't support
symlinks, which may or may not be an acceptable reason.