boost@1.89.0.bcr.1 (#6441)
There are changes made in
https://github.com/bazelbuild/bazel-central-registry/pull/6179 for
`1.88.0.bcr.2` that generally won't get picked up if a new version of
boost is added again because of existing `1.89.0` versions since the
`add_boost_version.py` script will look to copy from the latest
available version assuming it's the most up to date with Bazel. This
change aims to bring forward all the improvements made into the latest
available versions so future updates are far easier (and I don't have to
be the one to do it).