Upgrade lanelet2 to support boost 1.87+
Lanelet2 needed to pull in some commits from upstream (currently
unreleased) in order to be compatible wit boost 1.87.

NOTE: The BOOST_VERSION we use is still 1.83.0 instead of the latest
version. This so that users who are using an older version of boost can
still use lanelet2 without being forced to update boost at the same
time.

The max_compatibility_level arg is used to enable this version of
lanelet2 to remain usable as future versions of boost are released.

For more discussion, see:
https://github.com/bazelbuild/bazel-central-registry/discussions/6511