Add 'pending' commands (except the git ones) to toybox. (#6597)
In particular, this includes the 'sh' command, which is very helpful in
a lot of situations.

The git commands are excluded because they rely on OpenSSL's SHA
implementation and would pull that in as a dependency, increasing the
module's footprint significantly.

The toybox repo doesn't use stable release URLs, unfortunately.

@bazel-io skip_check unstable_url

---------

Co-authored-by: Marcel <maleo@google.com>