DO NOT MERGE - Testing CI with boost.test
This PR is meant to run CI on just the boost.test module. In the PR
(https://github.com/bazelbuild/bazel-central-registry/pull/6530),
boost.test hangs forever and is blocking the build. This PR is
separating out just boost.test so I can debug why the asan tests are
hanging forever in CI, but not locally on my machine when I test.