zlib@1.3.1.bcr.8 (#6818)
genrule are hostile to cross-platform remote execution because they bake
the path to bash into the command. `copy_file` will use a symlink
instead which works nicely for us.

I'm a little unsure about bumping the bazel_compat to 7.x - there are
still users on 6.x but bzlmod on 6.x is an unusual combo and they can
always pin to the release before this one if they don't need this fix?
If we really care though, we could supply the BUILD file as a patch
instead of an overlay...

Since module diff looks a bit confusing, the only real change to the
BUILD file is the `copy_public_header_` targets