Add riscv64 support for xz (#6824)
Hi there

This PR is mainly to enable the use of xz on rv64. I copied from
[5.4.5.bcr.6](https://github.com/bazelbuild/bazel-central-registry/tree/main/modules/xz/5.4.5.bcr.6)
to create this repo [main branch](https://github.com/ffgan/bazel-xz),
made a simple
[modification](https://github.com/ffgan/bazel-xz/commit/cad0c92d228581b92d7d63620efc4d41dc2b8301)
on top of it, and got this [riscv64
branch](https://github.com/ffgan/bazel-xz/tree/riscv64). The content of
this riscv64 branch is exactly what's in this PR, and after testing, it
successfully completes the build on riscv64.

Regarding the build on riscv64, please see the
[results](https://github.com/ffgan/bazel-xz/actions/runs/20210088678/job/58014587839)
of this CI run. You can check out the workflow file inside, which uses
docker+qemu to complete the build on rv64.
That's all about this PR. 

---
Below are some of my personal thoughts, which may have inaccuracies. If
there are any issues, please feel free to @ me directly - I'd be happy
to correct them immediately.

What puzzles me is why the xz module doesn't have a separate repo for
maintenance. I noticed that the github repo we link to in the BCR is the
direct source code of xz. Considering that it's indeed inappropriate to
store bazel-related rules in xz, we shouldn't place downstream files in
the upstream project. I can understand this.

I also noticed that we have an overlay, which means it needs to be
placed in the same directory as the xz source code without overwriting
it. At the same time, xz upstream cannot sync well with our bazel rules.
However, I believe we still have solutions. This scenario is too common
for downstream projects.

What we need to maintain are the bazel rules - adding bazel rules for a
specific version of xz. More specifically, we need to place our bazel
rules in the xz source code of that version. Then the problem becomes
simple. We don't need to store the xz source code in the git repository,
only the bazel rules. During testing, we specify the xz version to pull,
place the xz source code in the same directory as our bazel rules, and
then proceed with testing. This means we need a separate repository for
storing bazel rules, containing only the bazel rules. Each time we just
modify the bazel rules inside and the referenced xz version. This way,
we won't need thousands of lines of commits every time we submit a PR to
the xz module in BCR, which would turn version management and PR review
into a disaster.

The simplest example is to manage bazel rules like rules_go does. Split
the xz bazel module into a separate repository, then we maintain the
bazel rules in that repository, and the BCR repository also points to
that repository. This way we can also add CI for our xz bazel rules.



---
**Other Info**
Co-authored by: nijincheng@iscas.ac.cn;

---------

Signed-off-by: ffgan <sudoemt@gmail.com>