cereal@1.3.2.bcr.2: add a filegroup for bundling headers (#6881)
We have some use cases where we want to bundle the headers, but the BCR
build file doesn't have the filegroup that we have in our custom BUILD
for cereal (pre-bazelmod).

I wonder if there's a better way to do this than adding this upstream
BUILD rule. Does it make sense to have a local override of the BUILD
file for this one rule? Is it sensible to declare a build rule in my
project that globs files in an external repository? These are questions
for which, it seems the fastest way to get an answer is to create this
pull request with the possibly wrong solution :)

Please and thank you!

Co-authored-by: Jeremy Kun <j2kun@users.noreply.github.com>