Adding fire@0.2.0 (#6994)
Fixing some things around dependencies (fire is now no longer pulling in
language rules) and fixing incompatible flags, too.

This changes the usage of this module significantly. There won't be much
consumers, but I wanted to be a good citizen and increase the level
according to the purpose of the flag.